third_party/WebKit/LayoutTests/vr/getVRDisplays_one_display.html

<!DOCTYPE html>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script src="resources/fake-vr-displays.js"></script>
<script src="file:///gen/layout_test_data/mojo/public/js/mojo_bindings.js"></script>
<script src="file:///gen/device/vr/public/mojom/vr_service.mojom.js"></script>
<script src="resources/mock-vr-service.js"></script>
<script src="resources/test-constants.js"></script>
<script>
let fakeDisplays = fakeVRDisplays();

vr_test( (t) => {
  return navigator.getVRDisplays().then( (displays) => {
    t.step( () => {
      assert_true(displays != null);
      assert_equals(displays.length, 1);
      let display = displays[0];
      assert_equals(display.displayName, 'Google, Inc. Daydream View');
      assert_true(display.capabilities.canPresent);
      assert_approx_equals(display.depthNear, 0.01, FLOAT_EPSILON);
      assert_approx_equals(display.depthFar, 10000.0, FLOAT_EPSILON);
    }, "getVRDisplays returned correct results");
  }, (err) => {
    t.step( () => {
      assert_unreached("getVRDisplays rejected");
    });
  }).then( () => {
    t.done();
  });
}, [fakeDisplays["Pixel"]],
"WebVR properly returns a single display");

</script>
